+++ components/frame/config.cpp
@@ -1,3 +1,3 @@
 #include "config.h"
 
-const float Config::goalFrameLength = 20.0f; // 20 ms -> 50 fps
+const double Config::goalFrameLength = 20.0; // 20 ms -> 50 fps
